+++ backend/src/riot/summoner/summoner.controller.ts
@@ -1,16 +1,17 @@
 import { Controller, Get, Param } from "@nestjs/common";
+
 import { SummonerService } from "./summoner.service";
 
-@Controller("summoner")
+@Controller("riot/summoner")
 export class SummonerController {
-  private summonerService: SummonerService;
+  private readonly summonerService: SummonerService;
 
   constructor(summonerService: SummonerService) {
     this.summonerService = summonerService;
   }
 
   @Get(":name")
-  async findSummoner(@Param() { name }: { name: string }) {
-    return this.summonerService.findSummoner({ name });
+  async findSummonerByName(@Param("name") name: string) {
+    return this.summonerService.findByName(name);
   }
 }
