+++ backend/src/riot/summoner/summoner.module.ts
@@ -1,12 +1,12 @@
 import { Module } from "@nestjs/common";
-import { HttpModule } from "@nestjs/axios";
 
 import { SummonerController } from "./summoner.controller";
+
+import { AccountService } from "../account/account.service";
 import { SummonerService } from "./summoner.service";
 
 @Module({
-  imports: [HttpModule.register({ timeout: 1000 })],
   controllers: [SummonerController],
-  providers: [SummonerService],
+  providers: [AccountService, SummonerService],
 })
 export class SummonerModule {}
